Moving action types into unified location
