+++ src/applications/claims-status/actions/types.js
@@ -1,0 +1,54 @@
+// Routing
+export const SET_LAST_PAGE = 'SET_LAST_PAGE';
+
+// Claims
+export const FETCH_CLAIMS_ERROR = 'FETCH_CLAIMS_ERROR';
+export const FETCH_CLAIMS_PENDING = 'FETCH_CLAIMS_PENDING';
+export const FETCH_CLAIMS_SUCCESS = 'FETCH_CLAIMS_SUCCESS';
+export const GET_CLAIM_DETAIL = 'GET_CLAIM_DETAIL';
+export const SET_CLAIM_DETAIL = 'SET_CLAIM_DETAIL';
+export const SET_CLAIMS_UNAVAILABLE = 'SET_CLAIMS_UNAVAILABLE';
+
+// File uploads
+export const RESET_UPLOADS = 'RESET_UPLOADS';
+export const ADD_FILE = 'ADD_FILE';
+export const REMOVE_FILE = 'REMOVE_FILE';
+export const SET_UPLOADING = 'SET_UPLOADING';
+export const SET_UPLOADER = 'SET_UPLOADER';
+export const DONE_UPLOADING = 'DONE_UPLOADING';
+export const SET_PROGRESS = 'SET_PROGRESS';
+export const SET_UPLOAD_ERROR = 'SET_UPLOAD_ERROR';
+export const CANCEL_UPLOAD = 'CANCEL_UPLOAD';
+export const UPDATE_FIELD = 'UPDATE_FIELD';
+export const SET_FIELDS_DIRTY = 'SET_FIELD_DIRTY';
+export const SET_UNAUTHORIZED = 'SET_UNAUTHORIZED';
+
+// Notifications
+export const CLEAR_ADDITIONAL_EVIDENCE_NOTIFICATION =
+  'CLEAR_ADDITIONAL_EVIDENCE_NOTIFICATION';
+export const SET_ADDITIONAL_EVIDENCE_NOTIFICATION =
+  'SET_ADDITIONAL_EVIDENCE_NOTIFICATION';
+
+export const CLEAR_NOTIFICATION = 'CLEAR_NOTIFICATION';
+export const SET_NOTIFICATION = 'SET_NOTIFICATION';
+
+// Decision request
+export const SUBMIT_DECISION_REQUEST = 'SUBMIT_DECISION_REQUEST';
+export const SET_DECISION_REQUESTED = 'SET_DECISION_REQUESTED';
+export const SET_DECISION_REQUEST_ERROR = 'SET_DECISION_REQUEST_ERROR';
+
+// Appeals
+export const FETCH_APPEALS_ERROR = 'FETCH_APPEALS_ERROR';
+export const FETCH_APPEALS_PENDING = 'FETCH_APPEALS_PENDING';
+export const FETCH_APPEALS_SUCCESS = 'FETCH_APPEALS_SUCCESS';
+
+// Appeals error statuses
+export const BACKEND_SERVICE_ERROR = 'BACKEND_SERVICE_ERROR';
+export const RECORD_NOT_FOUND_ERROR = 'RECORD_NOT_FOUND_ERROR';
+export const USER_FORBIDDEN_ERROR = 'USER_FORBIDDEN_ERROR';
+export const VALIDATION_ERROR = 'VALIDATION_ERROR';
+
+// STEM Claims
+export const FETCH_STEM_CLAIMS_ERROR = 'FETCH_STEM_CLAIMS_ERROR';
+export const FETCH_STEM_CLAIMS_PENDING = 'FETCH_STEM_CLAIMS_PENDING';
+export const FETCH_STEM_CLAIMS_SUCCESS = 'FETCH_STEM_CLAIMS_SUCCESS';
