+++ src/applications/claims-status/claims-status-entry.jsx
@@ -11,12 +11,11 @@ import startReactApp from 'platform/startup/react';
 import createCommonStore from 'platform/startup/store';
 import startSitewideComponents from 'platform/site-wide';
 
-import ClaimsStatusApp from './containers/ClaimsStatusApp.jsx';
-import routes from './routes.jsx';
-import reducer from './reducers';
+import { setLastPage } from './actions';
+import ClaimsStatusApp from './containers/ClaimsStatusApp';
 import manifest from './manifest.json';
-
-import { setLastPage } from './actions/index.jsx';
+import routes from './routes';
+import reducer from './reducers';
 
 window.appName = manifest.entryName;
